backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2

# HTTP Client for proxying
httpx>=0.26.0
aiohttp>=3.9.1

# Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2

# Rate limiting
slowapi>=0.1.9

# CORS
starlette>=0.35.1
